requirements.txt
Flask==3.0.3
mysql-connector-python
openai>=1.0.0
